auth complete.

+++ frontend/src/Components/workoutsDetails.js
@@ -1,11 +1,19 @@
 import { useWorkoutsContext } from "../hooks/useWorkoutsContext"
+import { useAuthContext } from "../hooks/useAuthContext";
 
 const WorkoutsDetails = ({ workout }) => {
     const { dispatch } = useWorkoutsContext()
+    const { user } = useAuthContext()
 
     const handleClick = async () => {
+      if(!user){
+        return
+      }
       const response = await fetch('http://localhost:4000/api/workouts/' + workout._id, {
-        method: 'DELETE'
+        method: 'DELETE',
+        header: {
+          'Authorization': `Bearer ${user.token}`
+        }
       })
       const json = await response.json()
